Add global layout listener workaround to keep configuration up-to-date
Adds an additional step to the global layout listener for
AndroidComposeView to check if the Configuration has changed. This is a
workaround for a platform bug where View.onConfigurationChanged may not
always be called. The global layout listener will be one frame late,
which isn't ideal, but it ensures that it won't be out of sync for an
indefinite period of time.

After the merging of aosp/2467340, this additional check should be
relatively cheap, as the case where the Configuration doesn't change
won't cause unnecessary recompositions.

Fixes: 247143459

Test: Manual. Running the compose:integration-tests:material-catalog on
a 10.1 tablet running API 32 without this change results in
LocalConfiguration.current not always being updated. After this change,
it is updated.

Change-Id: Ia8e78372351c5ded9bcad53ba95b445a9e212530