Make the sole selected route not unselectable
It causes the provider to enter in a bad state. We could consider fixing
it, but we'd have to define a reasonable thing to do when the last route
is unselected. Releasing the session would be an obvious answer, but
then that would match the stop casting button, so why allow unselection
in the first place.

Bug: b/205124386
Test: Manually using the sample routing app.
Change-Id: I3096f026992ba9556db1fb2dee6ccb98a309b98e