uiautomator: Update testapp to respect window offsets for UI bound checks
The tests checking visible bounds failed in desktop mode because
the test activity returned the coordinates of the test view relative
to the host window, not absolute screen coordinates.
In non-multiwindow it was okay using the window relative coordinates
as long as the activity opens to the size of the entire screen.
However, when in desktop mode where wnidows can be resized and
in freeform it should use the absolute screen coordinates to match
the one returned by UiObject#getVisibleBounds().

Test: ./gradlew :test:uiautomator:i:t:connectedAndroidTest (UiObject2Test, UiObjectTest)
Bug: 409617767
Change-Id: Ie2f9598e85a47a5ac083bc9ac20dc5071833b1a5