Implementing WASM/JS support in DataStore via `sessionStorage`.
This changelist introduces WebStorage APIs for datastore for creating a datastore in `sessionStorage`, which is a synchronous API whose data is temporary and isolated to a single browser tab. The change also introduces a WebInterProcessCoordinator for `sessionStorage` which essentially does not do any locking as it is single-threaded by nature.

The WebInterProcessCoordinator and WebStorage API have been set up in a way that will allow adding support for `localStorage` and `OPFS` in the future.

Bug: b/316376114
Test: WebStorageTest.kt
Relnote: Implement WASM/JS support in DataStore via `sessionStorage`.
Change-Id: I60fade372879dc5531a774d94056c78adeadeddf