Add MediaSessionCompat.getBitmapDimensionLimit()
MediaSession will scale the bitmaps if they are larger than its
mMaxBitmapSize, which is from `config_mediaMetadataBitmapMaxSize`.
See MediaSession.setMetadata().

By having an API to query this number, it gives an app a chance
to optimize, e.g. the app could request thumbnails/bitmaps that
are within this limit, and hence save the scale, as well as reduce
the numbers of bitmaps.

E.g. YouTube Music on pixel 7a requests many thumbnails of size
853x480, while MediaSession.mMaxBitmapSize is 840, this ends up
with a scale down bitmap that is almost the same size.

Bug: 419883670
Flag: NONE
Test: gradle media:media:assemble
Test: gradle media:media:test
Test: local test of youtube/youtube music with the change and logged
the results from this method, all correct

Change-Id: I35ca35f430d5ed9a4a7ed1b46800118a0bf60bf4