Introduces UserSubspace API, a user-centric Subspace that follows the user's body part with configurable following behaviors.
It allows a developer to easily place content which follows the user around. This can be something like subtitles which display directly in the user's field of view. Or it can be a reference sheet, for example, which is off to the side but it always nearby.

Added the following integration test in the Compose TestApp for demonstration xr/compose/integration-tests/testapp/src/main/kotlin/androidx/xr/compose/testapp/lockedsubspace/UserSubspaceActivity.kt

Relnote: "Added new UserSubspace API. It allows a
developer to easily place content which follows
the user around."

Bug: 407766226

Test: Tested on Moohan emulator and device.
Change-Id: I6eec67e499bf4d442b73b0b1d369fc3c736d175e