Improve curved text width computation
Text rendering leaves a space before (and after) the text, but
this is not taken into account by getTextBounds, which seems
to be based on pixels.
Updated the width computation of curved text to account for that.
This was specially noticeable when using a background below the
curved text.
Relnote: Improve curved text width computation.
Test: Manual, updating screenshots.
Bug: 446601899

Change-Id: I924a4a6794a0eaee4638f2892764a03b9c6f6f1e