Selectively enable new FacadeClassJvmName API lint issue
This issue checks that JvmName is used for file-facade classes to provide a name for Java clients, so it is enabled only for projects which target Java clients.

Suppress all existing errors in projects targeting Java. This also updates the type for :glance:glance-material to be only for kotlin consumers, since glance projects aren't meant to target Java.

Bug: 165356974
Test: ./gradlew checkApi

Change-Id: I1eb5fb544661ba08e4a27a742d2d951a4a89d4e5