Scroll to center of viewport during scroll capture.
If requested coordinates are not fully visible, scroll them to the
center in order to reduce the chances that floating content at the
top/bottom is included in output.

This aligns with the RecylerView implementation.

To test this, added a new long screenshot demo that includes floating
text at the top and bottom. Verified that the bug occurs with that demo
and that it no longer occurs with this fix.

Bug: 433495972
Test: Manual reproduction of bug and verification of test described
Test: Run ScrollCaptureTest
above.
Relnote: N/A

Change-Id: Id0f85db4f6d82cf8def67a3a42a112a67c28334a