Only stop foreground service if no further starts pending
Pass in the start id token to stopSelf which lets the OS know to not
actually stop the service if there is a pending service start. This
should be what we want anyway since there is logic to handle
re-initializing the service if a start comes in after a shutdown.

This also indirectly fixes an unexpected OS issue where attempting to
call startForegroundService after stopping the service leads to a
ForegroundServiceDidNotStartInTimeException. This is because
startForegroundService sets fgRequired = true on the ServiceRecord
which is being brought down. The OS then sees this and thinks that the
service never called startForeground properly after
startForegroundService was called and throws an exception.

By passing in the start token, we avoid this issue by stopping the
service from being brought down in the first place.

Also add a test case to the integration test app that reproduces this
scenario fairly consistently.

Bug: 432069314
Relnote: "Fix foreground service stopping when pending commands"
Test: Integration test app
Change-Id: Iae822d63398e81ea0f1810c4d3cb4e0dafd4e13f