Fixes issue where presses could be stuck when the main thread is blocked
In cases when the main thread is blocked, coroutines we launch to handle emitting presses / releases inside clickable could be dispatched after a press-release-press were processed. This was causing the first press to be forgotten about, so we never dispatched the corresponding release for it. In this case we need to make sure that we correctly account for the fact that the first release's coroutine may execute after the second press logic happens.

Fixes: b/414319919
Test: ClickableTest#interactionSource_scrollableContainer_fastSuccessivePressesAndReleases
Change-Id: I145615effa2f8b582d367187152e3fcbc473c255