Handle class file dependencies in BanInappropriateExperimentalUsage
Previously, when finding the maven coordinates for a dependency, this lint check assumed dependencies would be contained in jar files. However, for a compose desktop dependency, the associated files were a directory of class files instead of a jar. This finds the coordinates of dependencies when they come from class files.

Bug: 424469950
Test: unit test for getting maven coordinates in BanInappropriateExperimentalUsageTest, removed baseline entry for compose runtime project
Change-Id: Id21e687eac3a232482dcab6d2b459ccab8607648