Hide meaningless info on Transformation APIs panel
In the Transformation APIs test case, panels show some
info about the distance between different panels. There
are some fields, such as "Distance to xxx (dest units)"
and "Distance to xxx (meters)," showing on each different
panel, but it is meaningless when comparing the distance
with itself, and the calculation result is always 0, which
results in people misunderstanding. This patch adds logic
to hide this info if comparing the distance itself to avoid
further misconstruing.

Bug: 443088664
Test: On the Moohan device
Change-Id: I110f8e573afb26295fe31031d1be17dfb1974cfb