Add demo to show how to append a Text Block below common Material3 components
TextBlockDemo shows a text block below Button, CheckboxButton, SwitchButton and RadioButton. The semantics are given to a parent Column, which clears the semantics on the children. This has the effect of appearing to be a combined component for accessibility, whilst all the functionality is unchanged when operating without TalkBack.
Bug: 416131768
Test: N/A
Relnote: N/A

Change-Id: I11252d10e9606da4593a0a0b58cbd9b367ea89f2