Move retain to separate module
The Compose Runtime's retain feature does not need to be implemented in
the runtime module itself for correct operation. This CL moves all
retain APIs and implementations out of the runtime module and into a new
runtime-retain module.

The net current.txt's remain the same with retain APIs moving into the
`androidx.compose.runtime.retain` package. That and documentation are
the only things impacted by this move.

Test: Moved retain tests
Relnote: """
   retain has been separated into a separate module as
   `androidx.compose.runtime:runtime-retain`. It now lives in the
   `androidx.compose.runtime.retain` package and can be depended on
   independently of the rest of the runtime.

   This initial release includes support for retaining values across
   configuration changes. Support for navigation will be included in
   an upcoming navigation release.
   """

Change-Id: Ia9443c80ebc3b286f92000e5350c5830a5aef91b