Use PUBLICATION mode in read-only reflection of FrameworkSQLite
Using NONE could cause a NPE due to the UnsafeLazyImpl nulling the initializer even though another thread is about to use it. PUBLICATION is safer and still faster than full synchronization, which in this case is still good as we expect the value to be the same even if calculated again.

Bug: 444049518
Test: N/A
Change-Id: I0084e8ac4a63748cb1703ca4aa03bdee670b39db