Differentiate between camera open failures
During camera opening, the CameraGraph or CameraPipe may be closed or
shutdown. There is a possibility that, if a CameraPipe is immediately
shutdown during camera opening, we may close the camera state while
waiting for onOpened to be back. In these scenarios, while we did not
get an opened camera, no errors happened. Tweak
RetrieveActiveCameraResult to different between camera open failures -
those with an error code vs those without, and do not print out an error
log when there wasn't an error.

Bug: 441150962
Test: CameraX test app and tests
Change-Id: I0023034a23290282b5591d24e6cbdafd7cdac4c9