Special case executing PRAGMA journal_mode = wal
If the journal mode is changed via the PRAGMA the internal connection pool is not reconfigured therefore special handling must be done when the journal mode is changed between WAL and non-WAL modes to invoke the dedicated Android APIs enableWriteAheadLogging() / disableWriteAheadLogging().

Not reconfiguring means the database will stay in a single connection, limiting parallelism.

Due to the increased connections by this fix, this CL exposed a flaw in the Room SupportSQLite wrapper, specifically, getting the last inserted row id and rows changed need to be done in a transaction to guaranteed getting the values from the primary connection.

Bug: 444286035
Test: SupportSQLiteDriverTest
Change-Id: I9ef65ae6a30ee97e04f3fb98ba4aaa2c70d0fe7b