Separate Kotlin and bytecode APIs for some functions returning value classes
In some cases, a function which returns a value class type does not have a mangled name in bytecode. Previously, the signature entries for these functions used the value class return type, but for bytecode the return type is the inlined value type. This change separates the entries into kotlin and bytecode versions.

Bug: 445490139
Test: ./gradlew checkApi
Relnote: N/A

Change-Id: Ie353438eb5c6835f90557789766aea889c4aafe5