Improve generic handling in EntryProvider
Use generics to better separate the superclass
of the entire entryProvider DSL (T) from the
individual classes used for a specific entry (K).

By moving the entry methods to within the builder,
we can remove the EntryClassProvider and
EntryProvider data classes from the public API surface
as they are only implementation details.

Relnote: "Improved the handling of generics for the
`entryProvider` DSL. If you were previously importing
`androidx.navigation3.runtime.entry`, this is no longer
required."
BUG: n/a
Test: existing tests pass

Change-Id: I299fc79a5e1ade5136d8ff6e1184adb1820851fe