Rearrange public API surface.
This change hides away the `WebStorage` API and creates explicitly defined WebLocalStorage and WebSessionStorage APIs, so they can be directly used without requiring a storage type ENUM flag.

This change also refactors PreferenceDataStoreFactory.web.kt to use WebSessionStorage by default, as this function does not take in a storage and therefore is not scalable, as it has no way of letting the user select the type of Web storage they want to use.

Bug: b/316376114
Test: Should not break existing tests
Change-Id: I35fa259b5d5940d599e8dce0efe79d9a15886895