[acms-fixit] Decouple CameraStateAdapter from UseCaseCameraConfig
This change removes the direct dependency on `CameraStateAdapter`
from `UseCaseCameraConfig` and `UseCaseGraphConfig`.

Previously, the `CameraStateAdapter` was passed through these
configuration objects down to the `UseCaseCamera`. This created
an unnecessary coupling, making the configuration objects aware
of an implementation detail of the camera's state management.

Following the recent introduction of the `GraphStateListener` pattern
(implemented via `GraphStateBridge`), the responsibility for observing
`GraphState` and updating the `CameraStateAdapter` now resides solely
within the `UseCaseManager`.

Bug: 437261143
Test: All Robolectric tests under integration layer.
Change-Id: I339d7fc93626b7bdeb868fab894e06bba590988d