Reset history to empty when no handlers active
Update `updateEnabledHandlerInfo` to accept a nullable handler and
add a check for when the `activeHandler` is null.

If no handlers are active, the processor will now correctly emit
a default, empty `NavigationEventHistory()` object.

This fixes a bug where the global `history` flow would get stuck
displaying the state of the last-known handler after all
handlers had been removed or disabled.

This fix allows the previously-ignored
`historyState_onHandlerChanges_updatesHistoryStack` test to
pass, so the `@Ignore` and TODO comments are removed.

Bug: 444734264
Test: NavigationEventDispatcherTest.kt
Change-Id: I12b3754e927057a1a32ff9a99a76f23df2bc1f3d