KGP 2.2.20 ABI files
KGP 2.2.20 includes a fix for the visibility of stable prop values in ABIs making them internal. These were present for Compose related libraries and now need to be ignored since they are being "removed", but they were never supposed to be part of the public API surface.

Relnote: N/A
Test: ./gradlew checkAbiRelease
Change-Id: I7430069d9c5439830c8e97d849c4e53e5b74ca0c