[Tooltip] Fix issue with tooltip caret not changing directions when scrolling
Change TooltipBox to trigger recomposition whenever the tooltip popup switches sides
so that `Modifier.layoutCaret` is called and the direction of the caret is
updated.

To get correct position of the popup, need to set `clippingEnabled = false`
in BasicTooltipBox, and need to update some of the positioning logic
within `TooltipPositionProviderImpl`

Change `positionProvider` to use the `LocalWindow` container size
instead of windowSize passed in to the Popup container to make sure
`positionProvider` and `Modifier.layoutCaret` use the same window
dimensions to calculate the directions. Otherwise, the directions aren't
always in sync.

Test: Manual testing (rich and plain tooltip)
Bug: b/438875827

Change-Id: I5ad1ebd9f53e4ef8084aadb10c685a04fb9bafe8