Always save the previously focused item
Prior to this CL we would save the previous focused
child when the saveFocusedChild() API was called.
This was called by FocusRestorer to save focus when
focus left a restorable focus group.

This successfully restores focus when we revisit a
focus group, but does not restore focus when the
activity is recreated. This is because it relies on
clearFocus to trigger a onExit, which in turn makes
a call to saveFocusedChild(). When an activity is
recreated, saveInstanceState is called before focus
is cleared, so it does not capture the previously
focused item.

When the new isFocusRestorationEnabled feature flag is
enabled, we now call saveFocusedChild automatically whenever
focus changes, so the focus restorer does not have to
save the previously focused child, but just implement
restoration.

With this change with the flag enabled, the saveFocusedChild()
API becomes a no-op, as focus is always saved, so we will be
able to deprecate that API as it is no longer needed once
the feature flag is rolled out.

Bug: 399936939
Bug: 324902166
Bug: 330111948
Test: Added new FocusRestorerConfigurationChangeTest

Change-Id: I7928a3714924f7c69988cbc95e09f2803864947a