Add SequencedExecutor for ordered task execution.
Implement SequencedExecutor, a wrapper for an arbitrary java.util.concurrent.Executor. It ensures that tasks submitted via `execute` are run in the order they were posted.

Employ a cooperative dispatch strategy, after executing a task it resubmits a new worker to the
underlying Executor to handle the next task in the queue.

Add tests for SequencedExecutor class.

Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroidTest
--info
-Pandroid.testInstrumentationRunnerArguments.class=androidx.javascriptengine.SequencedExecutorTest
Relnote: """Add `androidx.javascriptengine.common.SequencedExecutor`
class for ordered task execution. Tasks run in the same order they
were posted to the executor.
	"""
Bug: 435619571

Change-Id: I8f1d0a7e37c3f720ee2bbf68d5f1f9c3aca6a40b