[Concurrent camera composition] Support Effect in concurrent camera composition mode.
Effect is now supported in concurrent camera composition mode
(Two SingleCameraConfig instance with the same use cases)

The changes include
1) Adding the effect processing to the dual stream handling in
  StreamSharing
2) Added tests to ConcurrentCameraTest
3) Ensuring mirroring is handled appropriately

Please note that mirrorMode in VideoCapture will be ignored when
effect is set in concurrent camera composition mode.

Relnote: "Support CameraEffect in concurrent camera composition mode.
The effect will be applied on the composition output. Please note that
mirrorMode in VideoCapture will be ignored when effect is set in
concurrent camera composition mode."
Bug: 425565129
Test: ConcurrentCameraTest

Change-Id: If3d00d1463fe07a983f28d4355f822c0679db9bb