Fix ComposeCameraAppTest failures
Root cause: The original flow was to launch or switch the activity first and then set the expected screen and stream state settings to monitor it. With the flow, it couldn't guarantee that the STREAMING state would come after the expected screen and stream state have been set up.

Solution: Update the flow to set the expected screen and stream state settings first. Then, launch or switch the activity. Finally, wait for the stream state being received.

Bug: 446534379
Test: ComposeCameraAppTest
Change-Id: I65dff45644348dbc33e021ef893310bf661fe4d6