CacheWindow refill update
When re-filling the cache window we were using old data, so we sometimes avoided re-caching items that have changed. When the dataset has changed and we need to re-fill the window, we need to clean-up the cached information (prefetched and item sizes) so the algorithm for prefetching will run for all new items.

Fixes: b/445242486
Test: Added new test.
Change-Id: I6288c9c96bedc05aecf106e53d8cd83deee46c4e