Keep track of the measured stack item heights for the top, next, and next-next items.
The measured heights will be used for transition animations.

The CL also updates how StackState's topItem and topItemOffsetFraction
are computed to align them with the stack behaviour, where topItem is
only updated when the item completely moves off the viewport (when
scrolling forward) or when the previous item enters the viewport (when
scrolling backward).

Bug: 413429531
Test: StackStateTest, VerticalStackTest
Relnote: """Annotated StackState.topItemOffsetFraction with
    @FrequentlyChangingValue."""
Change-Id: Ib9e41b3fa8218c55fc607d9cc563ac72823ccdbf