Add recursive dispose and getChildren() to Entity
Moved the recursive disposal logic for child entities from the Java `BaseEntity` in `scenecore-spatial-core` to the Kotlin `BaseEntity` in `scenecore`.

This change introduces a new `getChildren()` method on the `Entity` interface, which returns a list of the entity's children.

The new disposal logic correctly avoids disposing of `MainPanelEntity` as part of a recursive call, as its lifecycle is tied to the session.

RelNote: Added `getChildren()` to Entity interface. Updated MainPanelEntity lifecycle.
Test: Unit tests.
Bug: b/362296608

Change-Id: Ia69d6c7e5cdb7b8fd269f4dddf72cdbc67b34ad5