Deprecate XRW allowlist API and clean up code
This API was introduced as a mitigation against the deprecation origin
trial for the X-Requested-With header in WebView. The trial has ended,
and the removal of the header has been cancelled, so this API no longer
serves any purpose.

This CL marks the API as deprecated and removes the implementation and
corresponding tests. The implementation of the API has also been removed
from WebView version 142.

RelNote: """Removal of the X-Requested-With header in WebView has been
    cancelled, and the header will again be sent on all requests by
    default. The allowlist API is no longer necessary, and has been
    deprecated."""
Fixes: 447094379
Test: N/A
Change-Id: I884c3108ca9632273953fe34e800ec2299d5b6ac