Add API MediaSessionCompat.getScaledBitmap()
A bitmap used in MediaMetadata and/or MediaDescription could
be scaled by some size limit (see MediaSession.setMetadata()).

And the same bitmap could be converted to ashmem explicitly or
implicitly multiple times when it's:

  * being used both in MediaMetadata and MediaDescription, where
    the bitmap will be converted to ashmem explicitly in
    MediaDescription, but not in MediaMetadata
  * being used multiple times, e.g. whenenver any MediaMetadata
    info is updated (but with the same bitmap), such as duration,
    and the same bitmap will be parcelled multiple times before
    the updated MediaMedata is sent to binder calls

This API provides apps a way to get an optimal bitmap that is

  * pre-scaled down to the size limit imposed in Media library
  * a shareable ashmem bitmap so it can be reused without being
    converted explicitly when write to parcel

For bitmaps with size already within the limit, and/or SDK_VERSION
doesn't support `Bitmap.asShared()`, the original bitmap will be
returned.

Bug: 419883670
Flag: NONE
Test: gradle media:media:assemble
Test: gradle media:media:test
Test: local test of youtube/youtube music with the change and logged
the results from this method, all correct

Change-Id: I2e4f95d2a2c60b165aacd1bd18731d11bf765563