Stop configuring session within Anchorable API.
This change also updates the kDoc to clearly specify that it is the consumers duty to configure the Session and get proper permissions. Also if the session does not have proper configurations, the anchorable API will just return and do nothing.

Configuring the session with the required Mode is not the duty of a JXR API. It is the app's job to request the proper permissions and configuring the session with the required Mode.

Bug: 446919466
Test: tested in an app.
Change-Id: I49f71c94bb4fb8b28e4604b20f2c39c1957d8708