Add PerceptionSpace.getScenePoseFromPerceptionPose
In preparation for deprecating the SpatialUser SceneCore APIs, create a
method that returns a ScenePose based off of a pose within the
perception space. This will allow developers to create a ScenePose based
off of a pose retrieved using ARCore for Jetpack XR APIs.

Relnote: "Adds an API for retrieving a ScenePose that represents the
composition of the perception space origin pose and a pose with respect
to the perception space origin."

Bug: 446995378
Bug: 441118965
Test: On Moohan device
Change-Id: I5b50a282fee4e643bbb07487317ea749d216c808