Update baselines for Kotlin-only functions which are no longer linked to the previous definition
When tracking for some functions was updated to Kotlin-only, they are no longer exact type matches for the version in the previous signature file. Previously, the declarations were still linked because the erased types were the same, but the metalava change in the topic stops comparing by erased types for kotlin-only callables.

The difference for all but one of these is whether non-final varargs are listed as ... or [], filed b/447153877 to work on making that consistent.

Test: `./gradlew checkApi`, androidx_with_metalava target
Bug: 447153877
Bug: 445490139
Change-Id: I2034ce94e033d8775be06e820186ebc39e4ab2ea