Removing RememberComposeView from SpatialPanel Implementation and replacing it with AndroidViewPanel.
RememberComposeView got a lot of backlash after API+2 and it has been our goal to start moving Spatial Composables away from it. For Background:

Its intended purpose is to create a ComposeView that is applicable to the local context in order to handle propagating the compositional context and could
be saved in a remember state. This design of rememberComposeView() was functional in developing other Spatial UI components, however it is apparent now that
its definition does not follow best practices of the Compose runtime and has led to limitations being next to difficult or impossible to solve.

By utilizing AndroidViewPanel for our base implementation for SpatialPanel we resolve the following issues:

1) We should neither create nor manipulate a View by returning it from the Composition hierarchy
2) 2D Compose UI content should be created as a subcomposition in the XR content
3) The view's composition creation and disposal logic must not be decoupled from its presentation

More information provided in the following doc:
go/compose-xr:deprecating-remembercomposeview

Test: Manually tested on  TestApp where I confirm layout logic for Orbiters, SubspaceLayout, SpatialElevation, SpatialDialog, and SpatialPopUp are correct

Bug: 415130218

Change-Id: I7dee0bb9bb947de4f354071f85c6efad38f02655