Fixed test app crash in onDestroy and auto-clear some listeners
SessionExt.kt enforces that the activity not be destroyed when the scene is accessed from the session.
This prevents calling session.scene during `onDestroy()`

To work around this, we now auto remove/clear some of our listeners
when our scene reaches the end of its lifecycle and when the
 mainPanelEntity is disposed.

We should do the same for other listeners ultimately tied to the
scene so that they do not need to be manually removed by devs.

Bug: 444686170
Bug: 447652647
Relnote: "All listeners in Scene and MainPanelEntity are
 automatically removed at the end of the Scene's lifecycle and
 when the MainPanelEntity is disposed, respectively."
Test: Ran all of the activities in the testapp and recreated them
Change-Id: I0cbf4d06bd202b7bbef516dd38e071155b03fb71