Reduce swiping iteration count in exercises
Prior to this change, the evaluation of a single test profile for
TimePickerBenchmark took over 15 minutes to run, and over 17 minutes for
DatePickerBenchmark.

This reduces the total swiping iteration count from 20 to reduce the
total time it costs to run.

Anecdotal change in running time:
- TimePickerBenchmarkTest: 15m 36s -> 5m 20s
- DatePickerBenchmarkTest: 17m 0s  -> 6m 23s

The benchmark exercise itself is iterated, so this reduction of loops
within the exercise itself does not eliminate the statistical accuracy
gained from iteration.

Bug: 444153437
Test: TimePickerBenchmarkTest
Test: DatePickerBenchmarkTest
Change-Id: I023f8b835c9af753fe1d9aafaa2fe74d7c937a7d