Refactor ViewModelStoreNavEntryDecorator
Refactored from a function into a class since it's functionally a factory for NavEntryDecorator. Renamed shouldRemoveViewModelStore to removeViewModelStoreOnPop to clarify that this lambda is relevant only when a wrapped entry is popped from the backStack.

Made the removeViewModelStoreOnPopCallback public so that users can wrap the default callback.

Bug: 444447434
Test: existing tests pass
Relnote: "Refactored ViewModelStoreNavEntryDecorator from a function to a class to better reflect its functionality as a factory for NavEntryDecorator, and renamed the decorator's shouldRemoveViewModelStore parameter to removeViewModelStoreOnPop to clarify that this callback is invoked only when an entry is popped from the backStack."
Change-Id: Iefdc582a213cfd177fd26862239980afd8f2de93